Dockerfiles/flink-connectors/connectors/requirements.txt
backoff
dataclasses-json
psycopg2-binary
requests
pyyaml